% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci}
\alias{calculate_dci}
\title{Calculate DCI for a \code{\link{river_net}} object}
\usage{
calculate_dci(net, form = NULL)
}
\arguments{
\item{net}{A \code{\link{river_net}} object, the river network over which the dendritic connectivity index will be calculated.}

\item{form}{A character string, either "potamodromous" or "diadromous" specifying the form of the DCI to be calculated.}
}
\value{
A \code{\link{data.frame}} of segment-level DCI values.
}
\description{
Calculate DCI for a \code{\link{river_net}} object
}
