% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci_dia}
\alias{calculate_dci_dia}
\title{Calculate non-thresholded diadromous DCI}
\usage{
calculate_dci_dia(
  all_members,
  net_nodes,
  seg_weights,
  outlet_seg,
  n.cores,
  quiet
)
}
\arguments{
\item{all_members}{An integer vector holding all assigned membership labels
in the \code{\link{river_net}} object.}

\item{net_nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}}
object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either
weighted or unweighted depending on parameters.}

\item{outlet_seg}{An integer indicating the membership label of the outlet segment}

\item{n.cores}{Optional integer specifying the number of cores to use for computation.
Defaults to 1. Parallelization is currently supported only on macOS and Linux.}

\item{quiet}{Logical. If \code{FALSE}, prints the global DCI and a plot of river segments
to the console. Defaults to \code{TRUE}.}
}
\value{
A data frame which holds raw and relative DCI scores for each segment.
}
\description{
Calculate non-thresholded diadromous DCI
}
\keyword{internal}
