% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci_pot_thresh}
\alias{calculate_dci_pot_thresh}
\title{Calculate non-thresholded invasive DCI}
\usage{
calculate_dci_pot_thresh(
  net,
  all_members,
  net_nodes,
  seg_weights,
  weighted,
  threshold,
  totweight,
  n.cores,
  quiet
)
}
\arguments{
\item{net}{A \link{river_net} object.}

\item{all_members}{An integer vector holding all assigned membership labels
in the \code{\link{river_net}} object.}

\item{net_nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}}
object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either
weighted or unweighted depending on parameters.}

\item{threshold}{Optional numeric value specifying a dispersal limit in map units.
If \code{NULL} (default), no limit is applied.}

\item{n.cores}{An optional integer value indicating the number of cores to
use. Defaults to 1. Currently only works on MacOS and Linux.}

\item{quiet}{Logical. If \code{FALSE}, prints the global DCI and a plot of river segments
to the console. Defaults to \code{TRUE}.}
}
\value{
A data frame which holds raw and relative DCI scores for each segment.
}
\description{
Calculate non-thresholded invasive DCI
}
\keyword{internal}
