% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_dendritic.R
\name{correct_divergences}
\alias{correct_divergences}
\title{Correct river divergences}
\usage{
correct_divergences(net, correct = TRUE)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{correct}{A logical value, when \code{FALSE}, the default, corrections are not automatically applied and a \code{\link{sf}} object of lines is returned with topological errors indicated. If \code{TRUE} errors are automatically corrected.}
}
\value{
If correct is \code{TRUE} a \code{\link{river_net}} object with the shorter of each divergent pair removed. If correct is \code{FALSE} a \code{\link{}} object with divergent pairs identified with a shared number in the new "divergent" column.
}
\description{
Correct river divergences
}
\keyword{internal}
