% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dci-package.R
\docType{package}
\name{dci-package}
\alias{dci}
\alias{dci-package}
\title{dci: Calculate the Dendritic Connectivity in River Networks}
\description{
Calculate and analyze ecological connectivity across the watercourse of river networks using the Dendritic Connectivity Index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aarkilanian/dci}
  \item Report bugs at \url{https://github.com/aarkilanian/dci/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
