% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_dendritic.R
\name{enforce_dendritic}
\alias{enforce_dendritic}
\title{Enforce dendritic river topology}
\usage{
enforce_dendritic(rivers, correct = FALSE)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{correct}{A logical value, when \code{FALSE}, the default, corrections are not automatically applied and a \code{\link[sf]{sf}} object of lines is returned with topological errors indicated. If \code{TRUE} errors are automatically corrected.}
}
\value{
If \code{correct} is \code{FALSE}, a \code{\link[sf]{sf}} object with non-dendritic topology indicated in columns "divergent" and "complex". These error columns indicate for each river line if that river is part of a divergent pair or complex confluence. The columns are populated by integers which indicate with which river they share a topological error. If \code{correct} is \code{TRUE}, a \code{rivers} object with automatic topological corrections applied is returned.
}
\description{
This function provides an interface to tools which can correct non-dendritic geometries in a river network. This tool can correct divergent rivers and complex confluences.
}
\details{
Divergent rivers are parts of a river network where a single tributary splits into two after a confluence. In a dendritic network small upstream rivers can only combine at confluences into a single river.

Complex confluences occur when confluences have over 2 input tributaries. In a dendritic network two tributaries only must combine into one at confluences.

If errors are being corrected manually, all divergent pairs must be reduced to only one river and complex confluences modified such that only 2 rivers join together. When corrections are done automatically the shorted divergent stream is kept.
}
\examples{
\dontrun{enforce_dendritic(rivers = sf_line_object)}
\dontrun{enforce_dendritic(rivers = sf_line_object, correct = TRUE)}
}
