% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{gather_dist}
\alias{gather_dist}
\title{Gather total distance between two nodes}
\usage{
gather_dist(from, to, nodes)
}
\arguments{
\item{from}{The from node's label}

\item{to}{The to node's label}

\item{nodes}{An \code{sf} object of the nodes of the \code{\link{river_net}} object with river attributes joined.}
}
\value{
The distance, in map units, between the two given segments.
}
\description{
Gather total distance between two nodes
}
\keyword{internal}
