% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{gather_neighb}
\alias{gather_neighb}
\title{Gather total length of neighbourhood around given source node}
\usage{
gather_neighb(net, source, distance)
}
\arguments{
\item{distance}{}
}
\value{
A numeric value indicating the total length of the neihbourhood
}
\description{
Gather total length of neighbourhood around given source node
}
\keyword{internal}
