% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{gather_perm}
\alias{gather_perm}
\title{Gather permeability between two nodes}
\usage{
gather_perm(from, to, nodes)
}
\arguments{
\item{from}{The from node's label}

\item{to}{The to node's label}

\item{nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}} object with river attributes joined.}
}
\value{
The permeability from 0 to 1 between the given nodes.
}
\description{
Gather permeability between two nodes
}
\keyword{internal}
