% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_rivers}
\alias{import_rivers}
\title{Prepare rivers for \code{\link{rivnet}} connectivity analyses}
\usage{
import_rivers(path, weight = NULL)
}
\arguments{
\item{path}{A character string or \code{\link{sf}} object, the path to a shapefile of river lines or \code{\link{sf}} object of rivers.}

\item{weight}{An optional double vector, river weights ranging from 0 to 1. Set to NULL by default.}
}
\value{
Object of class rivers prepared for input to \code{\link{rivnet}}
}
\description{
Read and prepare geospatial river lines data for dci.
}
