% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_attributes}
\alias{join_attributes}
\title{Join variables from network nodes to \code{\link{river_net}} object.}
\usage{
join_attributes(net, nodes, tolerance = NULL)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{nodes}{An \code{\link[sf]{sf}} object of binded "barrier", "sink", or "other" points.}

\item{tolerance}{An integer value, the maximum distance in map units from rivers to given points. Set to NULL by default.}
}
\description{
Join variables from network nodes to \code{\link{river_net}} object.
}
\keyword{internal}
