% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_invasions}
\alias{join_invasions}
\title{Join invasions to river segments}
\usage{
join_invasions(net, invasions)
}
\arguments{
\item{invasions}{A \code{invasions} object returned by \code{\link{import_points}}. This data is optional}
}
\value{
A \code{\link{river_net}} object with the supplied invasions integrated to edges
}
\description{
Join invasions to river segments
}
\keyword{internal}
