% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{membership_labeler}
\alias{membership_labeler}
\title{Membership labeling function passed to \code{\link[tidygraph]{map_bfs}}}
\usage{
membership_labeler(node, parent, path, env, ...)
}
\arguments{
\item{node}{The index of the current node.}

\item{parent}{The index of the parent node.}

\item{path}{A list of previous results.}

\item{env}{A parent environment holding past assigned labels.}

\item{...}{other parameters.}
}
\value{
An integer representing the current node's segment membership.
}
\description{
Membership labeling function passed to \code{\link[tidygraph]{map_bfs}}
}
\keyword{internal}
