% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{membership_labeling}
\alias{membership_labeling}
\title{Label nodes with integer segment member ID}
\usage{
membership_labeling(net)
}
\arguments{
\item{net}{A \link{river_net} object.}
}
\value{
A \code{\link{river_net}} object with nodes assigned segment membership labels.
}
\description{
Label nodes with integer segment member ID
}
\keyword{internal}
