% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{node_labeler}
\alias{node_labeler}
\title{Node labeling function passed to \code{\link[tidygraph]{map_bfs}}}
\usage{
node_labeler(node, parent, path, env, ...)
}
\arguments{
\item{node}{The index of the current node.}

\item{parent}{The index of the parent node.}

\item{path}{A list of previous results.}

\item{env}{A parent environment holding past assigned labels.}

\item{...}{other parameters.}
}
\value{
The correct node label. Either TRUE or FALSE.
}
\description{
Node labeling function passed to \code{\link[tidygraph]{map_bfs}}
}
\keyword{internal}
