% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{node_labeling}
\alias{node_labeling}
\title{Label nodes with logical vector binary IDs}
\usage{
node_labeling(net)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}
}
\value{
A \code{\link{river_net}} object with nodes assigned binary topological labels.
}
\description{
Label nodes with logical vector binary IDs
}
\keyword{internal}
