% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{path_between}
\alias{path_between}
\title{Find the path between two nodes}
\usage{
path_between(s1, s2)
}
\arguments{
\item{s1}{The label, a logical vector, of the origin node}

\item{s2}{The label, a logical vector, of the destination node}
}
\value{
A list of logical vectors in order to move from the origin to the destination node.
}
\description{
Find the path between two nodes
}
\keyword{internal}
