% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{path_to_root}
\alias{path_to_root}
\title{Find the path to the root (sink) of the river network}
\usage{
path_to_root(seg)
}
\arguments{
\item{seg}{A node label represented by a logical vector.}
}
\value{
A list of logical vectors in order to move from the origin to the root (sink).
}
\description{
Find the path to the root (sink) of the river network
}
\keyword{internal}
