% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename sf geometry column}
\usage{
rename_geometry(x, name)
}
\arguments{
\item{x}{an \code{\link[sf]{sf}} object}

\item{name}{a new geometry column name}
}
\value{
the original \code{\link[sf]{sf}} object with a renamed geometry
column
}
\description{
Code provided by user Spacedman
https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r
}
\keyword{internal}
