% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_net.R
\name{river_net}
\alias{river_net}
\title{Create a \code{river_net} object}
\usage{
river_net(
  rivers,
  barriers,
  outlet,
  poi = NULL,
  check = TRUE,
  tolerance = NULL,
  force_nodes = FALSE
)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{barriers}{A \code{barriers} object returned by \code{\link{import_points}}.}

\item{outlet}{A \code{outlet} object returned by \code{\link{import_points}}.}

\item{poi}{A \code{poi} object (points of interest) returned by \code{\link{import_points}}. This data is optional.}

\item{check}{A logical value, if \code{TRUE}, the default, dendritic topology of the river network is enforced with \code{\link{enforce_dendritic}}.}

\item{tolerance}{A numeric value specifying the distance in map units that points should be snapped to rivers. Set to NULL by default.}

\item{force_nodes}{A logical value indicating whether river network construction should be forced leadingto points potentially being deleted from the network to conform to dendritic network structure. Set to \code{FALSE} by default.}
}
\value{
An object of class \code{\link{river_net}} representing the river network formed by the geospatial lines and points provided.
}
\description{
\code{river_net} is a geospatial network structure of river data. It extends the tidy representation of geospatial network data from the \code{\link[sfnetworks]{sfnetwork}} data structure.
}
\examples{
\dontrun{ river_net(rivers = river_data, barriers = barrier_data, outlet = outlet_data)}
\dontrun{ river_net(river_data, barrier_data, outlet_data, poi = extra_data)}
\dontrun{ river_net(river_data, barrier_data, outlet_data, tolerance = 15}
}
