% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_rivers_at_points}
\alias{split_rivers_at_points}
\title{Split river lines around point locations}
\usage{
split_rivers_at_points(rivers, pts, tolerance = NULL)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{pts}{An \code{\link[sf]{sf}} object, points at which to split river lines.}

\item{tolerance}{An integer value, the maximum distance in map units from rivers to given points. Set to NULL by default.}
}
\description{
Split river lines around point locations
}
\keyword{internal}
