% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamaska.R
\docType{data}
\name{yamaska_barriers}
\alias{yamaska_barriers}
\title{Barrier point features for the Yamaska watershed}
\format{
An \code{sf} object with 620 LINESTRING features:
\describe{
\item{pass_1}{the passability of nodes in the network from 0 (impassable) to 1}
\item{pass_2}{the binary passability of nodes in the network, 0 if a barrier and 1 otherwise}
\item{geometry}{the geometry list column of river \code{POINT} features}
}
}
\usage{
yamaska_barriers
}
\description{
An \code{sf} object of the \code{POINT} geometries that make up imagined barriers on the
Yamaska watershed of Southern Québec. Features are projected to CRS:32198.
}
\keyword{datasets}
