% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamaska.R
\docType{data}
\name{yamaska_outlet}
\alias{yamaska_outlet}
\title{Outlet point feature for the Yamaska watershed}
\format{
An \code{sf} object with 1 POINT feature:
\describe{
\item{geometry}{the geometry list column of the outlet \code{POINT} feature}
}
}
\source{
\url{https://www.donneesquebec.ca/recherche/dataset/grhq}

\url{https://www.donneesquebec.ca/recherche/dataset/structure}
}
\usage{
yamaska_outlet
}
\description{
An \code{sf} object of the \code{POINT} geometry of the outlet of the Yamaska watershed
of Southern Québec. The feature is projected to CRS:32198.
}
\keyword{datasets}
