% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamaska_poi.R
\docType{data}
\name{yamaska_poi}
\alias{yamaska_poi}
\title{Points of interest data for the Yamaska watershed}
\format{
An object of class \code{\link[sf]{sf}} objects representing 5 \code{POINT} geometries and 2 columns:
\describe{
\item{FID}{A unique integer ID}
\item{geometry}{the geometry list column}
...
}
}
\usage{
yamaska_poi
}
\description{
An \code{\link[sf]{sf}} object reprenting point of interest points on the Yamaska watershed
}
\keyword{datasets}
