% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamaska.R
\docType{data}
\name{yamaska_rivers}
\alias{yamaska_rivers}
\title{River line features for the Yamaska watershed}
\format{
An \code{sf} object with 620 features:
\describe{
\item{qual}{a simulated weighting based on habitat quality of features}
\item{riv_length}{length of edge features in meters}
\item{geometry}{the geometry list column of edge line features}
}
}
\source{
\url{https://www.donneesquebec.ca/recherche/dataset/grhq}

\url{https://www.donneesquebec.ca/recherche/dataset/structure}
}
\usage{
yamaska_rivers
}
\description{
An \code{sf} object of the \code{LINESTRING} geometries that make up the rivers of the
Yamaska watershed of Southern Québec. The rivers make up the edges of the
\code{yamaska_net} object. Features are projected to CRS:32198.
}
\keyword{datasets}
