% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci}
\alias{calculate_dci}
\title{Calculate different forms of the DCI for a \code{\link{river_net}} object}
\usage{
calculate_dci(
  net,
  form,
  pass = NULL,
  weight = NULL,
  threshold = NULL,
  sites = NULL
)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{form}{A string specifying the form of the DCI to calculate: either "potamodromous", "catadromous", or "all".}

\item{pass}{The name of a column in the nodes table of net which holds the numeric passability of nodes. If none is specified all barriers are automatically considered to have 0 passability.}

\item{weight}{The name of column in the edges tables of net which holds numeric weights to be applied to river lengths. If none is specified, the DCI is calculated only with river lengths.}

\item{threshold}{An optional numeric value specifying a dispersal limit in map units. If NULL, the default, no limit is considered.}

\item{sites}{The name of a type of nodes in the node table of net. If specified, DCI results will be calculated at these sites and returned. If not specified, the DCI results will be reported on the rivers. See details for more.}
}
\value{
A \code{\link{sf}} object of the rivers from the provided \code{\link{river_net}} object with new columns speciying the segmental DCI values at each river location. If sites is not \code{NULL}, a \code{\link{sf}} object of the site points with their associated DCI scores.
}
\description{
Calculate different forms of the DCI for a \code{\link{river_net}} object
}
