% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci_dia}
\alias{calculate_dci_dia}
\title{Calculate non-thresholded diadromous DCI}
\usage{
calculate_dci_dia(all_members, net_nodes, seg_weights)
}
\arguments{
\item{all_members}{An integer vector holding all assigned membership labels in the \code{\link{river_net}} object.}

\item{net_nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}} object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either weighted or unweighted depending on parameters.}
}
\value{
A data frame which holds raw and relative DCI scores for each segment.
}
\description{
Calculate non-thresholded diadromous DCI
}
\keyword{internal}
