% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci_dia_thresh}
\alias{calculate_dci_dia_thresh}
\title{Calculate thresholded diadromous DCI}
\usage{
calculate_dci_dia_thresh(
  all_members,
  net_nodes,
  seg_weights,
  weighted,
  threshold,
  totweight
)
}
\arguments{
\item{all_members}{An integer vector holding all assigned membership labels in the \code{\link{river_net}} object.}

\item{net_nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}} object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either weighted or unweighted depending on parameters.}

\item{weighted}{A logical value indicating whether river lengths in seg_weights are weighted.}

\item{threshold}{An optional numeric value specifying a dispersal limit in map units. If NULL, the default, no limit is considered.}

\item{totweight}{The total length or weighted length of the whole network.}
}
\description{
Calculate thresholded diadromous DCI
}
\keyword{internal}
