% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{calculate_dci_pot}
\alias{calculate_dci_pot}
\title{Calculate non-thresholded potamodromous DCI}
\usage{
calculate_dci_pot(all_members, net_nodes, seg_weights, n.cores, quiet)
}
\arguments{
\item{all_members}{An integer vector holding all assigned membership labels
in the \code{\link{river_net}} object.}

\item{net_nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}}
object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either
weighted or unweighted depending on parameters.}

\item{n.cores}{An optional integer value indicating the number of cores to
use. Defaults to 1. Currently only works on MacOS and Linux.}

\item{quiet}{A logical value indicating whether the global DCI and a map of
segments should be printed to the console. Defaults to TRUE.}
}
\value{
A data frame which holds raw and relative DCI scores for each segment.
}
\description{
Calculate non-thresholded potamodromous DCI
}
\keyword{internal}
