% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_dendritic.R
\name{correct_complex}
\alias{correct_complex}
\title{Correct complex confluences}
\usage{
correct_complex(net, correct = TRUE, quiet = FALSE)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{correct}{A logical value, when \code{FALSE}, the default, corrections are not automatically applied and a \code{\link[sf]{sf}} object of lines is returned with topological errors indicated. If \code{TRUE} errors are automatically corrected.}
}
\value{
If correct is \code{TRUE} a \code{\link[sf]{sf}} object of rivers with complex confluences separated into two closely located valid confluences. If correct is \code{FALSE} a \code{\link[sf]{sf}} object with rivers participating in complex confluences labeled with the same number ina  new "complex" column.
}
\description{
Correct complex confluences
}
\keyword{internal}
