% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_dendritic.R
\name{correct_divergences}
\alias{correct_divergences}
\title{Correct river divergences}
\usage{
correct_divergences(net, correct = TRUE, quiet = FALSE)
}
\arguments{
\item{net}{A \link{river_net} object.}

\item{correct}{Logical. If \code{FALSE} (default), no changes are made and
topological issues are identified only. If \code{TRUE}, issues are
automatically corrected.}

\item{quiet}{Logical. If \code{FALSE}, prints the global DCI and a plot of river segments
to the console. Defaults to \code{TRUE}.}
}
\value{
If correct is \code{TRUE} a \code{\link{river_net}} object with the shorter of each divergent pair removed. If correct is \code{FALSE} a \code{\link[sf]{sf}} object with divergent pairs identified with a shared number in the new "divergent" column.
}
\description{
Correct river divergences
}
\keyword{internal}
