% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dci-package.R
\docType{package}
\name{dci-package}
\alias{dci}
\alias{dci-package}
\title{dci: Calculate the Dendritic Connectivity Index in River Networks}
\description{
Calculate and analyze ecological connectivity across the watercourse of river networks using the Dendritic Connectivity Index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aarkilanian/dci}
  \item Report bugs at \url{https://github.com/aarkilanian/dci/issues}
}

}
\author{
\strong{Maintainer}: Alex Arkilanian \email{a.arkilanian@gmail.com} (\href{https://orcid.org/0000-0002-5427-2867}{ORCID})

}
\keyword{internal}
