% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_dendritic.R
\name{enforce_dendritic}
\alias{enforce_dendritic}
\title{Enforce dendritic river topology}
\usage{
enforce_dendritic(rivers, divergence = TRUE, complex = TRUE)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{divergence}{A logical value, when \code{TRUE}, the default, divergences are corrected.}

\item{complex}{A logical value, when \code{TRUE}, the default, complex confluences are corrected.}
}
\value{
If \code{export} is \code{FALSE}, a \code{rivers} object with corrected dendritic topology. If \code{export} is \code{TRUE}, a \code{rivers} object with divergent river pairs given shared \code{divID} and rivers participating in complex confluences are given shared \code{complexID}.
}
\description{
This function provides an interface to tools which can correct non-dendritic geometries in a river network. This tool can correct divergent rivers and complex confluences.
}
\details{
Divergent rivers are parts of a river network where a single tributary splits into two after a confluence. In a dendritic network small upstream rivers can only combine at confluences into a single river.

Complex confluences occur when confluences have over 2 input tributaries. In a dendritic network two tributaries only must combine into one at confluences.

If errors are being corrected manually, all divergent pairs must be reduced to only one river and complex confluences modified such that only 2 rivers join together.
}
