% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_dci}
\alias{export_dci}
\title{Export DCI results to spatial format}
\usage{
export_dci(net, results, type = "rivers")
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{results}{\code{dci_results} object or a list of objects of DCI results returned by \code{\link{calculate_dci}}.}

\item{type}{The element of the river network for which results should be reported. Can be either "rivers" or any of the nodes included in the \code{\link{river_net}} object such as "barriers" or "poi".}
}
\value{
A \code{\link[sf]{sf}} object with the corresponding DCI results joined to each feature.
}
\description{
Export DCI results to spatial format
}
\examples{
\dontrun{
export_dci(net = net_name, results = dci_results, type = "rivers")
export_dci(net = net_name, results = dci_results, type = "poi")
export_dci(net = net_name, results = dci_results, type = "barriers")
export_dci(net = net_name, results = list(a, b, c), type = "rivers")
}
}
