% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_dci}
\alias{export_dci}
\title{Export DCI Results to Spatial Format}
\usage{
export_dci(net, results, type = "rivers", relative = FALSE, quiet = TRUE)
}
\arguments{
\item{net}{A \link{river_net} object.}

\item{results}{A \code{dci_results} object, or a list of such objects, as returned by
\code{\link[=calculate_dci]{calculate_dci()}}.}

\item{type}{A character string specifying which component of the river network
the results should be exported for. Valid options are \code{"rivers"} (default),
or any of the node layers included in the \link{river_net} object, such as
\code{"barriers"} or \code{"poi"}.}

\item{relative}{A logical value indicating whether relative DCI values should
be returned in addition to raw values. Defaults to \code{FALSE}.}

\item{quiet}{Logical. If \code{FALSE}, prints the global DCI and a plot of river segments
to the console. Defaults to \code{TRUE}.}
}
\value{
An \link{sf} object containing the corresponding DCI results joined
to the selected network component. If multiple results are supplied, result
columns are appended by a number corresponding to the index of the
associated results.
}
\description{
Exports the output of \code{calculate_dci()} as a spatial object with DCI values joined
to the relevant features in the river network.
}
\examples{
\dontrun{
export_dci(net = net_name, results = dci_results, type = "rivers")
export_dci(net = net_name, results = dci_results, type = "poi")
export_dci(net = net_name, results = dci_results, type = "barriers")
export_dci(net = net_name, results = list(result_a, result_b), type = "rivers")
}
}
