% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dci.R
\name{gather_dci}
\alias{gather_dci}
\title{Calculate sub-segmental DCI component between a single pair of segments}
\usage{
gather_dci(
  net,
  form,
  from,
  to,
  distance,
  pass,
  nodes,
  seg_weights,
  threshold,
  totweight,
  weighted
)
}
\arguments{
\item{net}{A \code{\link{river_net}} object.}

\item{from}{The origin segment's membership label.}

\item{to}{The destination segment's membership label.}

\item{distance}{The distance, in map units, between the two segments as calculated by \code{\link{gather_dist}}.}

\item{pass}{The passability between the two segments as calculated by \code{\link{gather_perm}}.}

\item{nodes}{An \code{\link{sf}} object of the nodes of the \code{\link{river_net}} object with river attributes joined.}

\item{seg_weights}{A data frame of each segments total length. Either
weighted or unweighted depending on parameters.}

\item{threshold}{An optional numeric value specifying a dispersal limit in
map units. If NULL, the default, no limit is considered.}

\item{totweight}{The total length or weighted length of the whole network.}

\item{weighted}{A logical value indicating whether river lengths in
seg_weights are weighted.}
}
\value{
The sub-segmental DCI component between given pair of segments.
}
\description{
Calculate sub-segmental DCI component between a single pair of segments
}
\keyword{internal}
