% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_points}
\alias{import_points}
\title{Prepare point data for \code{\link{river_net}} connectivity analyses}
\usage{
import_points(pts, type)
}
\arguments{
\item{pts}{A character string or \code{\link{sf}} object, the path to a shapefile of points or \code{\link{sf}} object of points.}

\item{type}{A character string, either of “bars”, “out”, "poi", or “inv” specifying the type of point. Barriers are "bars", outlet is "out", points of interest are "poi", and invasive locations are "inv".}
}
\value{
Object of class barriers, outlet, or poi prepared for input to \code{\link{river_net}}
}
\description{
Read and prepare geospatial point data for dci.
}
\examples{
\dontrun{ import_points(pts = path_to_shapefile)}
\dontrun{ import_points(pts = sf_point_object)}
}
