% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_rivers}
\alias{import_rivers}
\title{Prepare rivers for \code{\link{river_net}} connectivity analyses}
\usage{
import_rivers(rivers, quiet = FALSE)
}
\arguments{
\item{rivers}{A character string or \code{\link{sf}} object, the path to a shapefile of river lines or \code{\link{sf}} object of rivers.}

\item{quiet}{A logical value, if \code{FALSE} edited rivers are plotted with the original. Defaults to \code{FALSE}.}
}
\value{
Object of class rivers prepared for input to \code{\link{river_net}}
}
\description{
Read and prepare geospatial river lines data for \code{\link{river_net}}. Only the largest fully connected component will be preserved, river lines taking part in secondary networks will be discarded.
}
\examples{
\dontrun{ import_rivers(rivers = path_to_shapefile)}
\dontrun{ import_rivers{rivers = sf_line_object}}
}
