% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_net.R
\name{river_net}
\alias{river_net}
\title{Create a \code{river_net} object}
\usage{
river_net(rivers, barriers, sinks = NULL, poi = NULL, check = TRUE)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{barriers}{A \code{barriers} object returned by \code{\link{import_points}}.}

\item{sinks}{A \code{sinks} object returned by \code{\link{import_points}}. This data is optional.}

\item{poi}{A \code{poi} object returned by \code{\link{import_points}}. This data is optional.}

\item{check}{A logical value, if \code{TRUE}, the default, dendritic topology of the river network is enforced with \code{\link{enforce_dendritic}}.}

\item{tolerance}{An integer value, the maximum snapping distance in map units of supplied points relative to river lines. Points outside this distance will be discarded. Defaults to 10 map units.}
}
\value{
An object of class \code{\link{river_net}} representing the river network formed by the geospatial lines and points provided.
}
\description{
\code{river_net} is a geospatial network structure of river data. It extends the tidy representation of geospatial network data from the \code{\link[sfnetworks]{sfnetwork}} data structure.
}
