% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_net.R
\name{river_net}
\alias{river_net}
\title{Create a \code{river_net} object}
\usage{
river_net(
  rivers,
  barriers,
  sink = NULL,
  poi = NULL,
  check = TRUE,
  tolerance = NULL
)
}
\arguments{
\item{rivers}{A \code{rivers} object returned by \code{\link{import_rivers}}.}

\item{barriers}{A \code{barriers} object returned by \code{\link{import_points}}.}

\item{sink}{A \code{sink} object returned by \code{\link{import_points}}. This data is optional.}

\item{poi}{A \code{poi} object (points of interest) returned by \code{\link{import_points}}. This data is optional.}

\item{check}{A logical value, if \code{TRUE}, the default, dendritic topology of the river network is enforced with \code{\link{enforce_dendritic}}.}

\item{tolerance}{A numeric value specifying the distance in map units that points should be snapped to rivers. Set to NULL by default.}
}
\value{
An object of class \code{\link{river_net}} representing the river network formed by the geospatial lines and points provided.
}
\description{
\code{river_net} is a geospatial network structure of river data. It extends the tidy representation of geospatial network data from the \code{\link[sfnetworks]{sfnetwork}} data structure.
}
