% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamaska_outlet.R
\docType{data}
\name{yamaska_outlet}
\alias{yamaska_outlet}
\title{Outlet data for the Yamaska watershed}
\format{
An object of class \code{\link[sf]{sf}} objects representing 1 \code{POINT} geometry and 2 columns:
\describe{
\item{FID}{A unique integer ID}
\item{geometry}{the geometry list column}
...
}
}
\usage{
yamaska_outlet
}
\description{
An \code{\link[sf]{sf}} object reprenting the outlet point of the Yamaska watershed
}
\keyword{datasets}
